Narrowphase: move MinkowskiDiff and support functions to separate file
For the ContactPatchSolver, I need to have access to the support functions in the
same efficient way as GJK. For this reason, I moved the
MinkowskiDifference and all the support functions to their own .h/.hxx files.
